Refactor UPF UFFD snapshot handling
